Changed: Re-organized the exports in the file-formats-api crate
